fix(android): l'agente andava in crash ogni 10 secondi (0.5.2, build 7)
NoSuchMethodError: put(String, float) in sendSnapshot. Due cause sovrapposte:

1. La frequenza dello schermo (Display.Mode.refreshRate) e' un Float, e l'org.json
   di Android non ha un overload per float. Ora si converte a Double.
2. Il progetto dichiarava implementation("org.json:json"), ma org.json su Android
   lo fornisce il sistema: si compilava contro una versione e a runtime ne veniva
   usata un'altra. E' la causa per cui l'errore non si vedeva in compilazione.
   Dipendenza rimossa: si usa quella di piattaforma.

Effetto: il servizio moriva a ogni snapshot (10s), quindi telefono sempre
offline e servizio di Accessibilita' segnalato come "non funziona correttamente".

+++ android/app/src/main/java/com/matteomigliore/pcmanager/AgentService.kt
@@ -232,7 +232,7 @@ class AgentService : Service() {
             .put("device", JSONObject()
                 .put("manufacturer", Build.MANUFACTURER).put("model", Build.MODEL)
                 .put("android", Build.VERSION.RELEASE).put("sdk", Build.VERSION.SDK_INT)
-                .put("refreshRateHz", display?.mode?.refreshRate ?: 0f))
+                .put("refreshRateHz", (display?.mode?.refreshRate ?: 0f).toDouble()))
             .put("system", JSONObject()
                 .put("ramPct", round1(ramPct))
                 .put("ramUsedBytes", mem.totalMem - mem.availMem)
